feat: Use table for remote trainers
