+++ application/ui/src/features/remote-trainers/remote-trainers-page.module.css
@@ -1,48 +1,36 @@
-.pageHeader {
-    margin-bottom: var(--spectrum-global-dimension-size-250);
-}
-
-.pageHeader h1 {
-    margin-bottom: var(--spectrum-global-dimension-size-50);
-}
-
-.layout {
-    display: grid;
-    grid-template-columns: minmax(22.5rem, 26.25rem) minmax(0, 1fr);
-    min-height: 35rem;
+.table {
     border: 1px solid var(--spectrum-global-color-gray-300);
     border-radius: var(--spectrum-global-dimension-size-75);
     overflow: hidden;
 }
 
-.listPane {
-    padding: var(--spectrum-global-dimension-size-200);
-    background: var(--spectrum-global-color-gray-100);
-    border-right: 1px solid var(--spectrum-global-color-gray-300);
+.tableHeader {
+    padding: var(--spectrum-global-dimension-size-150) var(--spectrum-global-dimension-size-200);
+    background: var(--spectrum-global-color-gray-200);
+    text-transform: uppercase;
+    font-size: var(--spectrum-global-dimension-font-size-50);
+}
+
+.expandedContent {
+    padding: var(--spectrum-global-dimension-size-300);
+    background: var(--spectrum-global-color-gray-75);
+    border-top: 1px solid var(--spectrum-global-color-gray-300);
 }
 
 .addButton {
     margin-bottom: var(--spectrum-global-dimension-size-100);
-}
 
-.trainerCard {
-    width: 100%;
-    height: auto;
-    margin-bottom: var(--spectrum-global-dimension-size-100);
-    padding: var(--spectrum-global-dimension-size-150);
-    text-align: left;
-    background: var(--spectrum-global-color-gray-200);
-    border: 1px solid var(--spectrum-global-color-gray-300);
-    border-radius: var(--spectrum-global-dimension-size-75);
-}
-
-.trainerCardSelected {
-    border-color: var(--energy-blue);
-    box-shadow: inset 3px 0 var(--energy-blue);
-}
-
-.trainerCardSelected .trainerName {
-    color: var(--energy-blue);
+    /*
+     * Spectrum's .spectrum-Button applies an asymmetric vertical padding
+     * (`size-50 - 1px` top / `size-50 + 1px` bottom) as an optical correction
+     * calibrated for its own `adobe-clean` typeface. That font is not loaded
+     * here, so the label renders in the `-apple-system` fallback, whose
+     * cap-height/ascent ratio differs — the correction then overshoots and
+     * lifts the content ~1px above the button's centre. Restoring the symmetric
+     * `size-50` padding lets the button's own `align-items: center` centre the
+     * content, matching every ActionButton/FieldButton in the app.
+     */
+    padding-block: var(--spectrum-global-dimension-size-50);
 }
 
 .trainerName {
@@ -53,16 +41,6 @@
 .url {
     overflow-wrap: anywhere;
     color: var(--spectrum-global-color-gray-700);
-}
-
-.connectionTag {
-    padding: 0.125rem 0.375rem;
-    color: var(--energy-blue);
-    border: 1px solid var(--energy-blue);
-    border-radius: var(--spectrum-global-dimension-size-50);
-    font-size: var(--spectrum-global-dimension-font-size-50);
-    font-weight: 700;
-    letter-spacing: 0.04em;
 }
 
 .deviceTag {
@@ -94,44 +72,7 @@
     font-size: var(--spectrum-global-dimension-font-size-50);
 }
 
-.detailPane,
-.emptyDetail {
-    padding: var(--spectrum-global-dimension-size-300);
-    background: var(--spectrum-global-color-gray-75);
-}
-
-.detailHeader h2 {
-    margin: 0;
-}
-
-.detailActions {
-    align-self: center;
-    padding: var(--spectrum-global-dimension-size-50);
-    background: var(--spectrum-global-color-gray-100);
-    border: 1px solid var(--spectrum-global-color-gray-300);
-    border-radius: var(--spectrum-global-dimension-size-75);
-    box-shadow: 0 1px 2px color-mix(in srgb, var(--spectrum-global-color-static-black) 12%, transparent);
-}
-
-.destructiveAction {
-    display: flex;
-    margin-left: var(--spectrum-global-dimension-size-50);
-    padding-left: var(--spectrum-global-dimension-size-50);
-    border-left: 1px solid var(--spectrum-global-color-gray-300);
-}
-
-.deleteAction {
-    color: var(--spectrum-global-color-static-red-600);
-    fill: var(--spectrum-global-color-static-red-600);
-}
-
-.deleteAction:hover {
-    color: var(--spectrum-global-color-red-700);
-    fill: var(--spectrum-global-color-red-700);
-}
-
 .detailSection {
-    margin-top: var(--spectrum-global-dimension-size-300);
     padding: var(--spectrum-global-dimension-size-200);
     background: var(--spectrum-global-color-gray-100);
     border: 1px solid var(--spectrum-global-color-gray-300);
@@ -223,39 +164,6 @@
     overflow-wrap: anywhere;
 }
 
-.callout {
-    margin-top: var(--spectrum-global-dimension-size-200);
-    padding: var(--spectrum-global-dimension-size-50) 0 var(--spectrum-global-dimension-size-50)
-        var(--spectrum-global-dimension-size-150);
-    color: var(--spectrum-global-color-gray-700);
-    border-left: 3px solid var(--energy-blue);
-    font-size: var(--spectrum-global-dimension-font-size-75);
-}
-
-.formIntro {
-    padding: var(--spectrum-global-dimension-size-150);
-    background: var(--spectrum-global-color-gray-100);
-    border: 1px solid var(--spectrum-global-color-gray-300);
-    border-radius: var(--spectrum-global-dimension-size-75);
-}
-
-.formIntroLabel,
-.formIntroDescription {
-    display: block;
-    color: var(--spectrum-global-color-gray-700);
-    font-size: var(--spectrum-global-dimension-font-size-75);
-}
-
-.connectionTypeOption {
-    display: inline-block;
-    margin: var(--spectrum-global-dimension-size-75) 0;
-    padding: var(--spectrum-global-dimension-size-75) var(--spectrum-global-dimension-size-150);
-    color: var(--spectrum-global-color-static-white);
-    background: var(--spectrum-global-color-blue-500);
-    border-radius: var(--spectrum-global-dimension-size-50);
-    font-weight: 600;
-}
-
 .emptyList {
     display: block;
     padding: var(--spectrum-global-dimension-size-150);
@@ -267,18 +175,18 @@
 }
 
 @media (max-width: 48rem) {
-    .layout {
-        grid-template-columns: 1fr;
+    .tableHeader,
+    .row {
+        grid-template-columns: auto 2fr 1fr auto !important;
     }
 
-    .listPane {
-        border-right: 0;
-        border-bottom: 1px solid var(--spectrum-global-color-gray-300);
-    }
-
-    .detailActions {
-        width: 100%;
-        justify-content: flex-start;
+    .tableHeader > :nth-child(4),
+    .tableHeader > :nth-child(5),
+    .tableHeader > :nth-child(6),
+    .row > :nth-child(4),
+    .row > :nth-child(5),
+    .row > :nth-child(6) {
+        display: none;
     }
 
     .checkRow {
