+++ application/ui/src/features/remote-trainers/remote-trainers-table/remote-trainers-table.module.css
@@ -1,0 +1,43 @@
+.tableHeader {
+    padding: var(--spectrum-global-dimension-size-150) var(--spectrum-global-dimension-size-200);
+    background: var(--spectrum-global-color-gray-200);
+    text-transform: uppercase;
+    font-size: var(--spectrum-global-dimension-font-size-50);
+}
+
+.row {
+    padding: var(--spectrum-global-dimension-size-150) var(--spectrum-global-dimension-size-200);
+    border-top: 1px solid var(--spectrum-global-color-gray-300);
+}
+
+.rowExpanded {
+    background: var(--spectrum-global-color-gray-75);
+    box-shadow: inset 3px 0 var(--energy-blue);
+}
+
+.disclosureButton {
+    justify-self: start;
+}
+
+.trainerName {
+    font-weight: 700;
+    font-size: var(--spectrum-global-dimension-font-size-200);
+}
+
+.cardMetaText {
+    color: var(--spectrum-global-color-gray-600);
+    font-size: var(--spectrum-global-dimension-font-size-50);
+}
+
+.actionsCell {
+    justify-self: end;
+}
+
+.trainerRow {
+    cursor: pointer;
+    background: var(--spectrum-global-color-gray-100);
+
+    &:hover {
+        background: var(--spectrum-global-color-gray-200);
+    }
+}
